I think I'm done with the Project Operator. I left a bunch of "//comments" in my codes if you guys need help with reading this skeleton that we must use.
The only thing that I didn't do yet is the key and treemap/index. 
For the key, it seems like it's already implemented for the Project
operator. 

Not sure what he wants for the treemap/index. If it is just creating an
index, then it shouldn't be much of a problem. If not, then let me know.